requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dj-database-url==2.2.0
Django==5.1
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-spectacular==0.27.2
environ==1.0
flake8==7.1.1
flake8-annotations==2.9.1
flake8-quotes==3.4.0
flake8-variables-names==0.0.5
fonttools==4.54.1
freezegun==1.5.1
frozenlist==1.4.1
idna==3.7
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
Markdown==3.6
matplotlib==3.9.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.1.2
packaging==24.1
pathspec==0.12.1
pep8-naming==0.13.2
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.1
psycopg==3.2.2
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pydotplus==2.0.2
pyflakes==3.2.0
PyJWT==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.11.1
